[Fix] Numbers in ejercicio3.cpp must be ints

+++ ejercicio3.cpp
@@ -20,7 +20,7 @@ int main(){
 
     // ----- Declarations ------
 
-    float first_number, second_number;
+    int first_number, second_number;
 
     // ----- Taking Numbers ------
